encode type on id so it passes thru store
