+++ bucket/bucket.go
@@ -8,11 +8,32 @@ import (
 	"sync"
 )
 
+type libratoAttrs struct {
+	Min   int    `json:"display_min"`
+	Units string `json:"display_units_long"`
+}
+
+// When submitting data to Librato, we need to coerce
+// our bucket representation into something their API
+// can handle. Because there is not a 1-1 parity
+// with the statistical functions that a bucket offers and
+// the types of data the Librato API accepts (e.g. Librato does-
+// not have support for p50, p95, p99) we need to expand
+// our bucket into a set of LibratoMetric(s).
+type LibratoMetric struct {
+	Name   string        `json:"name"`
+	Time   int64         `json:"measure_time"`
+	Val    float64       `json:"value"`
+	Source string        `json:"source,omitempty"`
+	User   string        `json:"-"`
+	Pass   string        `json:"-"`
+	Attr   *libratoAttrs `json:"attributes,omitempty"`
+}
+
 type Bucket struct {
 	sync.Mutex
 	Id *Id
 	Vals []float64
-	Emtr Emitter
 }
 
 // Adding bucket a to bucket b copies the vals of bucket b and
@@ -29,7 +50,44 @@ func (b *Bucket) Add(otherM *Bucket) {
 // Relies on the Emitter to determine which type of
 // metrics should be returned.
 func (b *Bucket) Metrics() []*LibratoMetric {
-	return b.Emtr(b)
+	switch b.Id.Type {
+	case "measurement":
+		return b.EmitMeasurements()
+	case "counter":
+		return b.EmitCounters()
+	case "sample":
+		return b.EmiteSamples()
+	default:
+		panic("Undefined bucket.Id type.")
+	}
+}
+
+// The standard emitter. All log data with `measure.foo` will
+// be mapped to the MeasureEmitter.
+func (b *Bucket) EmitMeasurements() []*LibratoMetric {
+	metrics := make([]*LibratoMetric, 9)
+	metrics[0] = b.Metric("min", b.Min())
+	metrics[1] = b.Metric("median", b.Median())
+	metrics[2] = b.Metric("p95", b.P95())
+	metrics[3] = b.Metric("p99", b.P99())
+	metrics[4] = b.Metric("max", b.Max())
+	metrics[5] = b.Metric("mean", b.Mean())
+	metrics[6] = b.Metric("sum", b.Sum())
+	metrics[7] = b.Metric("count", float64(b.Count()))
+	metrics[8] = b.Metric("last", b.Last())
+	return metrics
+}
+
+func (b *Bucket) EmitCounters() []*LibratoMetric {
+	metrics := make([]*LibratoMetric, 1)
+	metrics[0] = b.Metric("sum", b.Sum())
+	return metrics
+}
+
+func (b *Bucket) EmiteSamples() []*LibratoMetric {
+	metrics := make([]*LibratoMetric, 1)
+	metrics[0] = b.Metric("last", b.Last())
+	return metrics
 }
 
 func (b *Bucket) Metric(name string, val float64) *LibratoMetric {
