+++ main.go
@@ -32,6 +32,7 @@ func main() {
 	// Can be passed to other modules
 	// as an internal metrics channel.
 	mchan := metchan.New(cfg.Verbose, cfg.MetchanUrl)
+	mchan.Start()
 
 	// The store will be used by receivers and outlets.
 	var st store.Store
